requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
audioop-lts==0.2.2
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
comtypes==1.4.16
distro==1.9.0
docopt==0.6.2
filelock==3.25.2
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.7.2
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
Js2Py==0.74
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.3
openai==2.29.0
packaging==26.0
pillow==12.1.1
pipwin==0.5.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyjsparser==2.7.1
pypiwin32==223
PyPrind==2.11.3
pySmartDL==1.3.4
python-dotenv==1.2.2
pyttsx3==2.99
pywin32==311
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
rich==14.3.3
safetensors==0.7.0
scipy==1.17.1
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.5
soupsieve==2.8.3
SpeechRecognition==3.15.1
standard-aifc==3.13.0
standard-chunk==3.13.0
sympy==1.14.0
tokenizers==0.22.2
torch==2.10.0
torchaudio==2.10.0
torchvision==0.25.0
tqdm==4.67.3
transformers==5.3.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
